fix: photo_type function name and metadata loader test structure
