+++ topo_processor/metadata/metadata_loaders/metadata_loader_imagery_historic.py
@@ -18,7 +18,7 @@ from topo_processor.stac.linz_provider import LinzProviders
 from topo_processor.stac.providers import Providers
 from topo_processor.stac.store import get_collection, get_item
 from topo_processor.util import (
-    h_i_photo_type_to_linz_geospatial_type,
+    historical_imagery_photo_type_to_linz_geospatial_type,
     nzt_datetime_to_utc_datetime,
     quarterdate_to_date_string,
     remove_empty_strings,
@@ -274,4 +274,4 @@ class MetadataLoaderImageryHistoric(MetadataLoader):
 
     def add_linz_geospatial_type(self, item: Item, photo_type) -> None:
 
-        item.linz_geospatial_type = h_i_photo_type_to_linz_geospatial_type(photo_type)
+        item.linz_geospatial_type = historical_imagery_photo_type_to_linz_geospatial_type(photo_type)
