+++ topo_processor/stac/item.py
@@ -29,15 +29,15 @@ class Item(Validity):
         self.id = item_id
         self.datetime = None
         self.properties = {
-                # TODO: [TDE-237] to generate release versioning
-                "processing:software": {"Topo Processor": "0.1.0"},
-                # TODO: decision to be made on version ref comments [TDE-230] hardcode to '1' for now
-                "version": "1",
+            # TODO: [TDE-237] to generate release versioning
+            "processing:software": {"Topo Processor": "0.1.0"},
+            # TODO: decision to be made on version ref comments [TDE-230] hardcode to '1' for now
+            "version": "1",
         }
         self.stac_extensions = set([StacExtensions.file.value])
         self.collection = None
         self.geometry_poly = None
-        self.linz_geospatial_type = None
+        self.linz_geospatial_type = ""
         self.assets = []
         self.schema = DefaultSchemaUriMap().get_object_schema_uri(pystac.STACObjectType.ITEM, pystac.get_stac_version())
