+++ topo_processor/util/__init__.py
@@ -1,7 +1,7 @@
 from .checksum import multihash_as_hex
 from .configuration import configuration
 from .conversions import (
-    h_i_photo_type_to_linz_geospatial_type,
+    historical_imagery_photo_type_to_linz_geospatial_type,
     nzt_datetime_to_utc_datetime,
     quarterdate_to_date_string,
     remove_empty_strings,
