+++ topo_processor/util/tests/conversions_test.py
@@ -1,7 +1,7 @@
 import pytest
 
 from topo_processor.util import (
-    h_i_photo_type_to_linz_geospatial_type,
+    historical_imagery_photo_type_to_linz_geospatial_type,
     nzt_datetime_to_utc_datetime,
     quarterdate_to_date_string,
 )
@@ -27,11 +27,11 @@ def test_quarter_date_to_utc_incorrect_format():
     assert returned_string == "nzam_pilot"
 
 
-def test_h_i_photo_type_to_linz_geospatial_type_empty_string():
-    returned_string = h_i_photo_type_to_linz_geospatial_type("")
+def test_historical_imagery_photo_type_to_linz_geospatial_type_empty_string():
+    returned_string = historical_imagery_photo_type_to_linz_geospatial_type("")
     assert returned_string == ""
 
 
-def test_h_i_photo_type_to_linz_geospatial_type_whitespace_case():
-    returned_string = h_i_photo_type_to_linz_geospatial_type(" B&w IR    ")
+def test_historical_imagery_photo_type_to_linz_geospatial_type_whitespace_case():
+    returned_string = historical_imagery_photo_type_to_linz_geospatial_type(" B&w IR    ")
     assert returned_string == "black and white infrared image"
